Move attachments to be a local resource in useForm()
